Do not throw errors when addProduct returns 400
A one-line change to add the `400` error code to the excepted codes when calling `mailchimp_ecommerce_add_product()`. This is because we now check for this error code, which is returned as the result of a product (or variant) already existing. Therefore, having a successful `updateProductVariant()` also throw an error is confusing to the end user.